refactor(core): extract symbol script helpers from runners
