+++ context.md
@@ -228,6 +228,7 @@ Current scope:
 - decompile can optionally emit `ghidra_apply_symbols.py` (`--emit-ghidra-script`) to apply recovered names as labels/functions inside Ghidra analysis sessions
 - decompile can optionally emit `ida_apply_symbols.py` (`--emit-ida-script`) to apply recovered names and pool-load comments inside IDA sessions
 - emitted Ghidra scripts also include pool-load comments derived from `pool[...]` annotations and recovered string hints, improving patching context directly in Ghidra
+- core pipeline implementation now keeps script-generation helpers in `runners_scripts.rs` (instead of inlining everything in `runners.rs`), reducing core runner file size and isolating RE-tool emit logic
 - decompile now also supports package-level scoping via repeatable `--app-package <name>`, so researchers can isolate pseudocode to selected app Dart packages and exclude unknown/dependency/framework noise more aggressively
 - report output now includes detected app package frequency (`function_scope.app_package_counts_top`) to guide package scoping without guesswork
 - report output now includes `function_scope.priority_package_hints`, the effective package hints applied to capped prioritization
