+++ crates/flutterdec-core/src/pipeline/runners_scripts.rs
@@ -1,0 +1,194 @@
+use std::fmt::Write as _;
+
+#[derive(Debug, Default, Clone, Copy)]
+struct SymbolScriptStats {
+    symbol_count: usize,
+    comment_count: usize,
+}
+
+fn parse_pool_annotation_index(annotation: &str) -> Option<u64> {
+    let raw = annotation.trim();
+    let body = raw.strip_prefix("pool[")?.strip_suffix(']')?;
+    body.parse::<u64>().ok()
+}
+
+fn pool_comment_literal(value: &str) -> String {
+    let trimmed = value.trim();
+    if trimmed.is_empty() {
+        return String::new();
+    }
+    let mut out = String::new();
+    for ch in trimmed.chars().take(160) {
+        match ch {
+            '\n' | '\r' | '\t' => out.push(' '),
+            _ => out.push(ch),
+        }
+    }
+    out
+}
+
+fn collect_ghidra_pool_comments(
+    disasm: &[FunctionDisassembly],
+    pool_value_hints: &HashMap<u64, String>,
+) -> Vec<(u64, String)> {
+    let mut comments: HashMap<u64, String> = HashMap::new();
+    for func in disasm {
+        for ins in &func.instructions {
+            let Some(pool_idx) = parse_pool_annotation_index(&ins.annotation) else {
+                continue;
+            };
+            let mut text = format!("pool[{pool_idx}]");
+            if let Some(value) = pool_value_hints.get(&pool_idx) {
+                let literal = pool_comment_literal(value);
+                if !literal.is_empty() {
+                    text.push_str(" = ");
+                    text.push_str(&literal);
+                }
+            }
+            comments.entry(ins.va).or_insert(text);
+        }
+    }
+    let mut out = comments.into_iter().collect::<Vec<_>>();
+    out.sort_by_key(|(va, _)| *va);
+    out
+}
+
+fn build_ghidra_symbol_script(
+    symbol_names: &HashMap<u64, String>,
+    pool_comments: &[(u64, String)],
+) -> Result<String> {
+    let mut entries = symbol_names
+        .iter()
+        .map(|(va, name)| (*va, name.clone()))
+        .collect::<Vec<_>>();
+    entries.sort_by_key(|(va, _)| *va);
+
+    let mut out = String::new();
+    out.push_str("# @category flutterdec\n");
+    out.push_str("# Generated by flutterdec\n\n");
+    out.push_str("from ghidra.program.model.symbol import SourceType\n");
+    out.push_str("import re\n\n");
+    out.push_str("def _sanitize_name(raw_name):\n");
+    out.push_str("    value = re.sub(r\"[^0-9A-Za-z_.$]\", \"_\", raw_name)\n");
+    out.push_str("    if not value:\n");
+    out.push_str("        return \"flutterdec_sym\"\n");
+    out.push_str("    if value[0].isdigit():\n");
+    out.push_str("        value = \"_\" + value\n");
+    out.push_str("    return value\n\n");
+    out.push_str("SYMBOLS = [\n");
+    for (va, name) in &entries {
+        let literal = serde_json::to_string(name).context("encode ghidra symbol literal")?;
+        writeln!(&mut out, "    (0x{va:x}, {literal}),")
+            .context("write ghidra symbol entry")?;
+    }
+    out.push_str("]\n");
+    out.push_str("POOL_COMMENTS = [\n");
+    for (va, comment) in pool_comments {
+        let literal = serde_json::to_string(comment).context("encode ghidra comment literal")?;
+        writeln!(&mut out, "    (0x{va:x}, {literal}),")
+            .context("write ghidra comment entry")?;
+    }
+    out.push_str("]\n\n");
+    out.push_str("for va, raw_name in SYMBOLS:\n");
+    out.push_str("    name = _sanitize_name(raw_name)\n");
+    out.push_str("    addr = toAddr(va)\n");
+    out.push_str("    fn = getFunctionAt(addr)\n");
+    out.push_str("    if fn is not None:\n");
+    out.push_str("        try:\n");
+    out.push_str("            fn.setName(name, SourceType.USER_DEFINED)\n");
+    out.push_str("        except Exception:\n");
+    out.push_str("            try:\n");
+    out.push_str("                createLabel(addr, name, True)\n");
+    out.push_str("            except Exception:\n");
+    out.push_str("                pass\n");
+    out.push_str("        continue\n");
+    out.push_str("    try:\n");
+    out.push_str("        createLabel(addr, name, True)\n");
+    out.push_str("    except Exception:\n");
+    out.push_str("        pass\n");
+    out.push('\n');
+    out.push_str("for va, text in POOL_COMMENTS:\n");
+    out.push_str("    addr = toAddr(va)\n");
+    out.push_str("    try:\n");
+    out.push_str("        setEOLComment(addr, text)\n");
+    out.push_str("    except Exception:\n");
+    out.push_str("        pass\n");
+    Ok(out)
+}
+
+fn write_ghidra_symbol_script(
+    path: &Path,
+    symbol_names: &HashMap<u64, String>,
+    pool_comments: &[(u64, String)],
+) -> Result<SymbolScriptStats> {
+    let script = build_ghidra_symbol_script(symbol_names, pool_comments)?;
+    fs::write(path, script).with_context(|| format!("write ghidra script: {}", path.display()))?;
+    Ok(SymbolScriptStats {
+        symbol_count: symbol_names.len(),
+        comment_count: pool_comments.len(),
+    })
+}
+
+fn build_ida_symbol_script(
+    symbol_names: &HashMap<u64, String>,
+    pool_comments: &[(u64, String)],
+) -> Result<String> {
+    let mut entries = symbol_names
+        .iter()
+        .map(|(va, name)| (*va, name.clone()))
+        .collect::<Vec<_>>();
+    entries.sort_by_key(|(va, _)| *va);
+
+    let mut out = String::new();
+    out.push_str("# Generated by flutterdec\n\n");
+    out.push_str("import ida_funcs\n");
+    out.push_str("import ida_name\n");
+    out.push_str("import idc\n");
+    out.push_str("import re\n\n");
+    out.push_str("def _sanitize_name(raw_name):\n");
+    out.push_str("    value = re.sub(r\"[^0-9A-Za-z_.$]\", \"_\", raw_name)\n");
+    out.push_str("    if not value:\n");
+    out.push_str("        return \"flutterdec_sym\"\n");
+    out.push_str("    if value[0].isdigit():\n");
+    out.push_str("        value = \"_\" + value\n");
+    out.push_str("    return value\n\n");
+    out.push_str("SYMBOLS = [\n");
+    for (va, name) in &entries {
+        let literal = serde_json::to_string(name).context("encode ida symbol literal")?;
+        writeln!(&mut out, "    (0x{va:x}, {literal}),").context("write ida symbol entry")?;
+    }
+    out.push_str("]\n");
+    out.push_str("POOL_COMMENTS = [\n");
+    for (va, comment) in pool_comments {
+        let literal = serde_json::to_string(comment).context("encode ida comment literal")?;
+        writeln!(&mut out, "    (0x{va:x}, {literal}),").context("write ida comment entry")?;
+    }
+    out.push_str("]\n\n");
+    out.push_str("NAME_FLAGS = ida_name.SN_NOWARN | ida_name.SN_NOCHECK\n\n");
+    out.push_str("for va, raw_name in SYMBOLS:\n");
+    out.push_str("    name = _sanitize_name(raw_name)\n");
+    out.push_str("    if not ida_name.set_name(va, name, NAME_FLAGS):\n");
+    out.push_str("        try:\n");
+    out.push_str("            idc.create_insn(va)\n");
+    out.push_str("            ida_funcs.add_func(va)\n");
+    out.push_str("        except Exception:\n");
+    out.push_str("            pass\n");
+    out.push_str("        ida_name.set_name(va, name, NAME_FLAGS)\n");
+    out.push('\n');
+    out.push_str("for va, text in POOL_COMMENTS:\n");
+    out.push_str("    idc.set_cmt(va, text, 0)\n");
+    Ok(out)
+}
+
+fn write_ida_symbol_script(
+    path: &Path,
+    symbol_names: &HashMap<u64, String>,
+    pool_comments: &[(u64, String)],
+) -> Result<SymbolScriptStats> {
+    let script = build_ida_symbol_script(symbol_names, pool_comments)?;
+    fs::write(path, script).with_context(|| format!("write ida script: {}", path.display()))?;
+    Ok(SymbolScriptStats {
+        symbol_count: symbol_names.len(),
+        comment_count: pool_comments.len(),
+    })
+}
